Improve color and fontSize animations
